requirements.txt
numpy
scikit_image
Shapely
tqdm
albumentations
labelme
Pillow
skimage
